refactor: refactor map code
