+++ components/map/SymbolList.tsx
@@ -1,11 +1,11 @@
-import { Pressable, ScrollView } from 'native-base';
+import { ScrollView } from 'native-base';
 import React from 'react';
 
+import PressableSymbol from './PressableSymbol';
 import CustomRectangle from './symbols/CustomRectangle';
 import { useAppStore } from '../../stores/store';
 
 function SymbolList() {
-  const selectedSymbol = useAppStore((state) => state.selectedSymbol);
   const setSelectedSymbol = useAppStore((state) => state.setSelectedSymbol);
 
   const handleSymbolPress = (symbolName: string) => {
@@ -14,22 +14,12 @@ function SymbolList() {
 
   return (
     <ScrollView flexDirection="column">
-      <Pressable
-        padding={3}
-        bgColor={selectedSymbol === 'test' ? 'amber.300' : undefined}
-        alignItems="center"
-        onPress={() => handleSymbolPress('test')}
-      >
+      <PressableSymbol onPress={() => handleSymbolPress('test')}>
         <CustomRectangle size={{ height: 30, width: 30 }} color="#f00" strokeStyle />
-      </Pressable>
-      <Pressable
-        padding={3}
-        bgColor={selectedSymbol === 'test1' ? 'amber.300' : undefined}
-        alignItems="center"
-        onPress={() => handleSymbolPress('test1')}
-      >
+      </PressableSymbol>
+      <PressableSymbol onPress={() => handleSymbolPress('test1')}>
         <CustomRectangle size={{ height: 30, width: 30 }} color="#2D3ED3" />
-      </Pressable>
+      </PressableSymbol>
     </ScrollView>
   );
 }
